test: add setupIntersectionObserverMock for mocking IntersectionObserver in tests
