+++ src/tests/components/IconItem.test.ts
@@ -6,6 +6,7 @@ import * as _iconPreview from "../../models/CustomizableIcon";
 
 import IconItem from "../../components/icons/IconItem.svelte";
 import type { Icon } from "../../models/Icon";
+import { setupIntersectionObserverMock } from "../utils/interceptionObserverHelper";
 
 // Mock the stores
 vi.mock("../stores/icon-customizations.store", () => ({
@@ -13,47 +14,6 @@ vi.mock("../stores/icon-customizations.store", () => ({
 		selectIcon: vi.fn()
 	}
 }));
-
-/**
- * Utility function that mocks the `IntersectionObserver` API. Necessary for components that rely
- * on it, otherwise the tests will crash. Recommended to execute inside `beforeEach`.
- * @param intersectionObserverMock - Parameter that is sent to the `Object.defineProperty`
- * overwrite method. `jest.fn()` mock functions can be passed here if the goal is to not only
- * mock the intersection observer, but its methods.
- *
- * Code from https://stackoverflow.com/a/58651649/7851973
- */
-export function setupIntersectionObserverMock({
-	root = null,
-	rootMargin = "",
-	thresholds = [],
-	disconnect = () => null,
-	observe = () => null,
-	takeRecords = () => [],
-	unobserve = () => null
-} = {}): void {
-	class MockIntersectionObserver implements IntersectionObserver {
-		readonly root: Element | null = root;
-		readonly rootMargin: string = rootMargin;
-		readonly thresholds: ReadonlyArray<number> = thresholds;
-		disconnect: () => void = disconnect;
-		observe: (target: Element) => void = observe;
-		takeRecords: () => IntersectionObserverEntry[] = takeRecords;
-		unobserve: (target: Element) => void = unobserve;
-	}
-
-	Object.defineProperty(window, "IntersectionObserver", {
-		writable: true,
-		configurable: true,
-		value: MockIntersectionObserver
-	});
-
-	Object.defineProperty(global, "IntersectionObserver", {
-		writable: true,
-		configurable: true,
-		value: MockIntersectionObserver
-	});
-}
 
 beforeEach(() => setupIntersectionObserverMock({ observe: vi.fn() }));
 
@@ -82,7 +42,7 @@ describe("IconItem Component", () => {
 		const button = getIconButton();
 
 		await fireEvent.click(button);
-		expect(customizedIcon.selectIcon).toHaveBeenCalledWith(mdiIcon);
+		expect(mockOnSelectIcon).toHaveBeenCalledWith(mdiIcon);
 	});
 
 	// it('lazy loads image for streamdeck origin', async () => {
